server/pip_requirements.txt
flask==0.10.1
jinja2==2.8
werkzeug==0.11.8
flask-sqlalchemy==2.1
pymysql==0.7.2
fabric3==1.11.1